pyproject.toml
[tool.poetry]
name = "hirzenbach-bot"
version = "0.1.0"
description = ""
authors = ["Jan Hohenheim <jan@hohenheim.ch>"]
license = "MIT"
readme = "README.md"
packages = [{include = "hirzenbach_bot"}]

[tool.poetry.dependencies]
python = "^3.9"
python-telegram-bot = "^20.0a4"
python-dotenv = "^0.21.0"
requests = "^2.28.1"
openai = "^0.23.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"
black = "^22.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
